old hx711

+++ R4_STM_CODE/Core/Inc/hx711.hh
@@ -1,7 +1,6 @@
 #ifndef HX711__LIB
 #define HX711__LIB
 #include <main.hh>
-#include <cstdlib>
 
 class HX711{
 	GPIO_TypeDef *Dt_gpio;
@@ -9,7 +8,7 @@ class HX711{
 	GPIO_TypeDef *Sck_gpio;
 	uint16_t Sck_pin;
 	int32_t OffsetInBits;
-	double BitsToGramRatio;
+	int32_t BitsToGramRatio;
 	uint8_t GAIN = 1;
 
 public:
@@ -17,9 +16,9 @@ public:
 		  GPIO_TypeDef *Sck_gpio, uint16_t Sck_pin);
 	int32_t getWeigthInGramsWithOffset(uint16_t times = 10);
 	int32_t getOffsetInBits() const;
-	double getBitsToGramRatio() const;
+	int32_t getBitsToGramRatio() const;
 	void setBitsToGramRatio(int32_t newCalibrationFactor);
-	void initialCalibration(double testLoadInGrams, uint16_t calibrationTime = 10000);
+	void initialCalibration(int32_t testLoadInGrams, uint16_t calibrationTime = 10000);
 	void tare();
 	void setGain(uint8_t gain = 128);
 	int32_t ReadValue();
